[FEATURE] Enhanced Frontend output settings
Two new options are introduced to the Frontend output FlexForm:

The single.activateLink setting disables the single view
If this checkbox is not set the the detail action will redirect
the user back to the list view.

settings.flexform.main.requiredFields allows the selection of fields
that must not be empty in a mail record. Otherwise it will not be
displayed. This allows the implementation of a select field where
the user can choose if he wants his data to be displayed.

Resolves: #57952
Releases: 2.1
Change-Id: I52beb8f912e357c8be53e6ac178a08a3d235d60a